# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WeMarket API",
    "version": "1.3.0",
    "description": "WeMarket QR Menu & Small Business Platform API"
  },
  "servers": [
    {
      "url": "https://wemarket.onrender.com/api",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "orderCapability": {
        "type": "apiKey",
        "in": "header",
        "name": "x-order-capability"
      }
    }
  },
...
  "paths": {
...
    "/api/export/store/{storeId}/excel/orders": {
      "get": {
        "tags": [
          "Export"
        ],
        "summary": "주문 내역 Excel 내보내기",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "storeId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "start_date",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "end_date",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Excel 파일 다운로드"
          }
        }
      }
    },
...
    "/api/kds/stores/{storeId}/orders": {
      "get": {
        "tags": [
          "KDS"
        ],
        "summary": "활성 조리중/대기중 주문 목록 조회",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "storeId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "KDS 주문 목록 (preparing, ready 상태)"
          }
        }
      }
    },
    "/api/kds/stores/{storeId}/orders/{orderId}/status": {
      "post": {
        "tags": [
          "KDS"
        ],
        "summary": "주문 조리 단계 상태 전이",
        "description": "preparing → ready → completed 흐름으로 상태 변경",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "storeId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "preparing",
                      "ready",
                      "completed"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "상태 전이 완료"
          }
        }
      }
    },
...
    "/api/orders": {
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "주문 생성",
        "description": "`Idempotency-Key` 헤더(UUID 권장)를 보내면 네트워크 타임아웃 후 재시도해도\n중복 주문이 생성되지 않는다. 동일 키의 이전 성공 응답이 그대로 재생되며\n응답에 `Idempotency-Replayed: true` 헤더가 포함된다.\n",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "중복 주문 방지용 고유 키 (권장)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "주문 생성 완료"
          },
          "409": {
            "description": "동일 Idempotency-Key 요청이 처리 중"
          },
          "422": {
            "description": "동일 Idempotency-Key로 다른 본문 전송"
          }
        }
      }
    },
    "/api/orders/{orderId}/customer-token": {
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "고객용 웹 푸시 토큰 등록 (KDS 취소 알림용)",
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "토큰 등록 완료"
          }
        }
      }
    },
    "/api/orders/customer/history": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "고객 본인의 주문 내역 조회",
        "parameters": [
          {
            "in": "query",
            "name": "phone",
            "schema": {
              "type": "string"
            },
            "description": "전화번호 기반 조회"
          },
          {
            "in": "query",
            "name": "toss_key",
            "schema": {
              "type": "string"
            },
            "description": "토스 키 기반 조회"
          }
        ],
        "responses": {
          "200": {
            "description": "주문 내역 목록"
          }
        }
      }
    },
    "/api/orders/store/{storeId}/detailed-stats": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "매장별 상세 통계 (그래프용)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "storeId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "시간대별/카테고리별 상세 통계"
          }
        }
      }
    },
    "/api/orders/store/{storeId}/stats": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "매장별 주문 통계 (상단 카드용)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "storeId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "오늘/이번 달 매출, 주문 수 등"
          }
        }
      }
    },
    "/api/orders/store/{storeId}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "매장별 주문 목록 조회",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "storeId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "주문 목록 (페이지네이션)"
          }
        }
      }
    },
    "/api/orders/{id}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "주문 단일 상세 조회",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "주문 상세 정보"
          },
          "404": {
            "description": "주문을 찾을 수 없음"
          }
        }
      },
      "delete": {
        "tags": [
          "Orders"
        ],
        "summary": "주문 삭제",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "orderCapability": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "주문 삭제 완료"
          }
        }
      }
    },
    "/api/orders/{id}/status": {
      "put": {
        "tags": [
          "Orders"
        ],
        "summary": "주문 상태 업데이트",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "상태 변경 완료"
          }
        }
      }
    },
    "/api/orders/{id}/cancel": {
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "주문 취소 (재고 복구 포함)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelOrderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "주문 취소 완료"
          }
        }
      }
    },
    "/api/orders/{id}/return-exchange": {
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "주문 반품/교환 신청",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReturnExchangeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "반품/교환 접수 완료"
          }
        }
      }
    },
    "/api/orders/store/{storeId}/eta": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "주문 예상 소요 시간(ETA) 조회",
        "parameters": [
          {
            "in": "path",
            "name": "storeId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "items",
            "schema": {
              "type": "string"
            },
            "description": "주문 아이템 배열(JSON 문자열) [{\"product_id\":1,\"quantity\":2}]"
          }
        ],
        "responses": {
          "200": {
            "description": "예상 소요 시간",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "etaMinutes": {
                      "type": "integer"
                    },
                    "activeOrdersAhead": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/search": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "주문 검색 (관리자용)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "storeId",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "paymentStatus",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "paymentMethod",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "주문 검색 결과"
          }
        }
      }
    },
...
    "/api/payments/order/{orderId}/cancel": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "주문 전체 취소",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cancelReason": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "취소 완료"
          }
        }
      }
    },
...
    "/api/sse/order/{orderId}": {
      "get": {
        "tags": [
          "SSE"
        ],
        "summary": "주문 상태 SSE 구독",
        "description": "주문의 실시간 상태 변경을 SSE로 수신합니다.",
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "주문 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "SSE 스트림 (text/event-stream)"
          }
        }
      }
    },
...
    "/api/v1/orders": {
      "get": {
        "tags": [
          "OpenAPI v1"
        ],
        "summary": "주문 내역 검색",
        "security": [
          {
            "apiKeyHeader": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "주문 목록"
          }
        }
      },
      "post": {
        "tags": [
          "OpenAPI v1"
        ],
        "summary": "외부 신규 주문 생성 (write 스코프 필수)",
        "security": [
          {
            "apiKeyHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "주문 생성 완료"
          }
        }
      }
    },
    "/api/v1/orders/{id}": {
      "get": {
        "tags": [
          "OpenAPI v1"
        ],
        "summary": "단일 주문 상세 조회",
        "security": [
          {
            "apiKeyHeader": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "주문 상세"
          }
        }
      }
    },
...
  }
}